init: add script pip-install.sh
One of the problem that
    pip install -r requirements.txt
is not working due to error:
    >>> import fitz
    Traceback (most recent call last):
      File "<stdin>", line 1, in <module>
      File "/venv/site-packages/fitz/__init__.py", line 1, in <module>
        from frontend import *
    ModuleNotFoundError: No module named 'frontend'

To fix error we need to install PyMuPDF just after fitz.
Bit it is impossible to do with pure pip.

So a small workaround script has been added to install python
packages one by one in a fixed order.